backend/django/requirements.txt
django==3.2
djangorestframework
django-cors-headers
psycopg2-binary
gunicorn
requests
django-dotenv
django-storages
boto3